python_service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdfminer.six==20221105
Werkzeug==3.0.1

